Aula: 'Trabalhando com o Status Code da aplicação'

+++ criando-aplicacoes-web-com-spring-web-mvc/projects/spring-web-mvc/src/main/java/me/gabreuw/springwebmvc/exception/JediNotFoundException.java
@@ -1,7 +1,7 @@
 package me.gabreuw.springwebmvc.exception;
 
 public class JediNotFoundException extends RuntimeException {
-    public JediNotFoundException(Integer id) {
+    public JediNotFoundException(Long id) {
         super("Não foi possível encontrar um Jedi com o id: " + id);
     }
 }
